# Maya SceneShape VP2 optimizations 