This improves performance of our VP2 drawing as tested on 2 production assets. The first is a digi-double with ~370k polys across 20 shapes. The second is a robot with ~300k polys across ~200 shapes.

Note all timings are taken on a cycle animation (either walk cycle or callisthenics test) after playing through once to get the filesystem IO out of the picture as much as possible. Also, they weren't measured too scientifically, just playing back and watching the FPS reported in the Viewport HUD.

| | Before | After |
:-: | :-: | :-: |
Human | 11 fps | 121 fps
Robot | 20 fps | 38 fps

A vtune profile taken after these changes shows Maya / GPU crosstalk as the primary bottleneck now, but I've also added a todo for the most prominent Cortex bit remaining. That'd take a big enough re-write of the traversal function that I wasn't sure I'd be able to get to it by weeks end, so thought it was worth putting this up now as I'm off work for the rest of April.

Its probably worth going through the code a commit at a time, as the logic of each changed is explained in each commit message, and the final diff is a bit much to look at in one go.